Cleaned up documentation, deleted obsolete files
