--- twitter-general/tweet-covid19-emoji/src/Test/Scala/EmojiUtilityTest.scala
@@ -1,3 +1,0 @@
-object EmojiUtilityTest {
-
-}
